backend/requirements.txt
Flask==2.3.3
boto3==1.34.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
uuid==1.30
marshmallow==3.20.1
marshmallow-dataclass==8.6.0
Pillow==10.0.1
requests==2.31.0
gunicorn==21.2.0
pydub==0.25.1 